backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.5.0",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "google-generativeai>=0.8.3",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.39.0",
]
